requirements.txt
Django==4.2.16
djangorestframework>=3.14.0
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.26
Pillow>=8.2.0,<8.3.0